v0.3.2.0
This version is just a variant of v0.3.1.0, which just relocates the static positions that is suitable for taskbar-less secondary phone-monitor. If you're enabling taskbar on the secondary monitor, simply switch back to v0.3.1.0 (although for future updates, this version will be used)

Checks.ahk is used for easier editing of code for checking for important events.